move KITZstack, SimpleZstack and Remote to own files
