+++ stp_zmq/remote.py
@@ -1,0 +1,150 @@
+from stp_core.common.config.util import getConfig
+import time
+import zmq
+from stp_core.common.log import getlogger
+import sys
+
+
+logger = getlogger()
+
+
+def set_keepalive(sock, config):
+    # This assumes the same TCP_KEEPALIVE configuration for all sockets which
+    # is not ideal but matches what we do in code
+    sock.setsockopt(zmq.TCP_KEEPALIVE, 1)
+    sock.setsockopt(zmq.TCP_KEEPALIVE_INTVL, config.KEEPALIVE_INTVL)
+    sock.setsockopt(zmq.TCP_KEEPALIVE_IDLE, config.KEEPALIVE_IDLE)
+    sock.setsockopt(zmq.TCP_KEEPALIVE_CNT, config.KEEPALIVE_CNT)
+
+
+class Remote:
+    def __init__(self, name, ha, verKey, publicKey, config=None):
+        # TODO, remove *args, **kwargs after removing test
+
+        # Every remote has a unique name per stack, the name can be the
+        # public key of the other end
+        self.name = name
+        self.ha = ha
+        # self.publicKey is the public key of the other end of the remote
+        self.publicKey = publicKey
+        # self.verKey is the verification key of the other end of the remote
+        self.verKey = verKey
+        self.socket = None
+        # TODO: A stack should have a monitor and it should identify remote
+        # by endpoint
+
+        self._numOfReconnects = 0
+        self._isConnected = False
+        self._lastConnectedAt = None
+        self.config = config or getConfig()
+
+        # Currently keeping uid field to resemble RAET RemoteEstate
+        self.uid = name
+
+    def __repr__(self):
+        return '{}:{}'.format(self.name, self.ha)
+
+    @property
+    def isConnected(self):
+        if not self._isConnected:
+            return False
+        lost = self.hasLostConnection
+        if lost:
+            self._isConnected = False
+            return False
+        return True
+
+    def setConnected(self):
+        self._numOfReconnects += 1
+        self._isConnected = True
+        self._lastConnectedAt = time.perf_counter()
+
+    def firstConnect(self):
+        return self._numOfReconnects == 0
+
+    def connect(self, context, localPubKey, localSecKey, typ=None):
+        typ = typ or zmq.DEALER
+        sock = context.socket(typ)
+        sock.curve_publickey = localPubKey
+        sock.curve_secretkey = localSecKey
+        sock.curve_serverkey = self.publicKey
+        sock.identity = localPubKey
+        set_keepalive(sock, self.config)
+        addr = 'tcp://{}:{}'.format(*self.ha)
+        sock.connect(addr)
+        self.socket = sock
+        logger.trace('connecting socket {} {} to remote {}'.
+                     format(self.socket.FD, self.socket.underlying, self))
+
+    def disconnect(self):
+        logger.debug('disconnecting remote {}'.format(self))
+        if self.socket:
+            logger.trace('disconnecting socket {} {}'.
+                         format(self.socket.FD, self.socket.underlying))
+
+            if self.socket._monitor_socket:
+                logger.trace('{} closing monitor socket'.format(self))
+                self.socket._monitor_socket.linger = 0
+                self.socket.monitor(None, 0)
+                self.socket._monitor_socket = None
+                # self.socket.disable_monitor()
+            self.socket.close(linger=0)
+            self.socket = None
+        else:
+            logger.debug('{} close was called on a null socket, maybe close is '
+                         'being called twice.'.format(self))
+
+        self._isConnected = False
+
+    @property
+    def hasLostConnection(self):
+
+        if self.socket is None:
+            logger.warning('Remote {} already disconnected'.format(self))
+            return False
+
+        events = self._lastSocketEvents()
+
+        if events:
+            logger.trace('Remote {} has monitor events: {}'.
+                         format(self, events))
+
+        # noinspection PyUnresolvedReferences
+        if zmq.EVENT_DISCONNECTED in events or zmq.EVENT_CLOSED in events:
+            logger.debug('{} found disconnected event on monitor'.format(self))
+
+            # Reverse events list since list has no builtin to get last index
+            events.reverse()
+
+            def eventIndex(eventName):
+                try:
+                    return events.index(eventName)
+                except ValueError:
+                    return sys.maxsize
+
+            connected = eventIndex(zmq.EVENT_CONNECTED)
+            delayed = eventIndex(zmq.EVENT_CONNECT_DELAYED)
+            disconnected = min(eventIndex(zmq.EVENT_DISCONNECTED),
+                               eventIndex(zmq.EVENT_CLOSED))
+            if disconnected < connected and disconnected < delayed:
+                return True
+
+        return False
+
+    def _lastSocketEvents(self, nonBlock=True):
+        return self._get_monitor_events(self.socket, nonBlock)
+
+    @staticmethod
+    def _get_monitor_events(socket, non_block=True):
+        monitor = socket.get_monitor_socket()
+        events = []
+        # noinspection PyUnresolvedReferences
+        flags = zmq.NOBLOCK if non_block else 0
+        while True:
+            try:
+                # noinspection PyUnresolvedReferences
+                message = recv_monitor_message(monitor, flags)
+                events.append(message['event'])
+            except zmq.Again:
+                break
+        return events
