+++ stp_zmq/test/test_zstack_communication.py
@@ -8,7 +8,7 @@ from stp_core.network.auth_mode import AuthMode
 from stp_core.test.helper import Printer, prepStacks, \
     checkStacksConnected, chkPrinted
 from stp_zmq.test.helper import genKeys
-from stp_zmq.zstack import KITZStack
+from stp_zmq.kit_zstack import KITZStack
 
 
 @pytest.mark.skip('Use a packet capture tool')
